# Lazily generate property getter 